Return observations dropped by simsum

dropbig.multisimsum returns observations dropped by simsum while computing summary statistics for a simulation study.

# S3 method for multisimsum
dropbig(x, ...)

Arguments

x
An object of class multisimsum.

...
Ignored.

Value

An object of class dropbig.multisimsum if dropbig was set to TRUE when calling multisimsum, NULL otherwise. It contains two slots: big_estvarname and big_se. Both slots consist in the portion of the original data that was identified to have estimated values or standard errors exceeding the limits defined by max and semax.

See also

Examples

data(frailty)
ms <- multisimsum(
  data = frailty, par = "par", true = c(
    trt = -0.50,
    fv = 0.75
  ), estvarname = "b", se = "se", methodvar = "model",
  by = "fv_dist", dropbig = TRUE, max = 6, semax = 3
)

#> `ref` was not specified, Cox, Gamma set as the reference

dropbig(ms)

#> Dropped point estimates:
#>   i        b        se par fv_dist             model
#>  79 2.097306 0.4200795  fv   Gamma   Cox, Log-Normal
#>  79 2.122200 0.4700095  fv   Gamma RP(P), Log-Normal
#>  Row number       SE        NA NA    NA              NA
#>          27 1.797547 0.4934171 fv Gamma Cox, Log-Normal
#>          93 1.406227 0.4929476 fv Gamma Cox, Log-Normal
#>         143 1.533038 0.5376231 fv Gamma Cox, Log-Normal
#>         243 1.343827 0.5046963 fv Gamma Cox, Log-Normal
#>         550 1.343645 0.5311375 fv Gamma Cox, Log-Normal
#>         904 1.590170 0.4826771 fv Gamma Cox, Log-Normal