Return observations dropped by simsum

dropbig.simsum returns observations dropped by simsum while computing summary statistics for a simulation study.

# S3 method for simsum
dropbig(x, ...)

Arguments

x
An object of class simsum.

...
Ignored.

Value

An object of class dropbig.simsum if dropbig was set to TRUE when calling simsum, NULL otherwise. It contains two slots: big_estvarname and big_se. Both slots consist in the portion of the original data that was identified to have estimated values or standard errors exceeding the limits defined by max and semax.

See also

Examples

data("MIsim")
x <- simsum(
  data = MIsim, estvarname = "b", true = 0.5, se = "se",
  methodvar = "method", mcse = TRUE,
  dropbig = TRUE, max = 3, semax = 1.5
)

#> `ref` was not specified, CC set as the reference

d <- dropbig(x)
d

#> Dropped point estimates:
#> # A tibble: 17 x 4
#>    dataset method         b    se
#>      <dbl> <chr>      <dbl> <dbl>
#>  1      47 CC       0.0801  0.130
#>  2     111 CC       0.970   0.191
#>  3     146 CC       0.944   0.171
#>  4     160 CC       0.927   0.164
#>  5     324 CC       0.944   0.190
#>  6     345 CC       1.00    0.178
#>  7     368 CC       0.928   0.172
#>  8     476 CC       0.974   0.170
#>  9     539 CC       0.0275  0.152
#> 10     539 MI_T     0.0703  0.133
#> 11     606 CC       0.988   0.159
#> 12     761 CC      -0.00924 0.164
#> 13     761 MI_T    -0.148   0.143
#> 14     761 MI_LOGT -0.108   0.159
#> 15     843 CC       0.984   0.175
#> 16     843 MI_T     0.976   0.182
#> 17     843 MI_LOGT  0.938   0.162
#> 
#> Dropped standard errors:
#> # A tibble: 8 x 4
#>   dataset method      b    se
#>     <dbl> <chr>   <dbl> <dbl>
#> 1      19 MI_T    0.460 0.216
#> 2     159 MI_T    0.860 0.228
#> 3     211 MI_T    0.464 0.215
#> 4     256 CC      0.278 0.219
#> 5     417 CC      0.601 0.210
#> 6     417 MI_LOGT 0.653 0.223
#> 7     551 MI_LOGT 0.543 0.227
#> 8     585 MI_LOGT 0.737 0.207