get_data.multisimsum

Extract data from a multisimsum object

# S3 method for multisimsum
get_data(x, sstat = "all", description = FALSE,
  ...)

Arguments

x
An object of class multisimsum.

sstat
Summary statistics to include; can be a scalar value or a vector. Possible choices are:

all, all the summary statistics are returned This is the default option.

nsim, the number of replications with non-missing point estimates and standard error.

thetamean, average point estimate.

thetamedian, median point estimate.

se2mean, average standard error.

se2median, median standard error.

bias, bias in point estimate.

empse, empirical standard error.

mse, mean squared error.

relprec, percentage gain in precision relative to the reference method.

modelse, model-based standard error.

relerror, relative percentage error in standard error.

cover, coverage of a nominal level% confidence interval.

bccover, bias corrected coverage of a nominal level% confidence interval.

power, power of a (1 - level)% level test.#'

description
Append a column to the returned data.frame with a description of each summary statistic? Defaults to FALSE.

...
Ignored.

Value

A data.frame containing summary statistics from a simulation study.

See also

Examples

data(frailty)
ms <- multisimsum(
  data = frailty, par = "par", true = c(
    trt = -0.50,
    fv = 0.75
  ), estvarname = "b", se = "se", methodvar = "model",
  by = "fv_dist"
)

#> `ref` was not specified, Cox, Gamma set as the reference

get_data(ms)

#>            stat           est         mcse             model    fv_dist par
#> 1          nsim  9.760000e+02           NA        Cox, Gamma      Gamma  fv
#> 2     thetamean  7.375961e-01           NA        Cox, Gamma      Gamma  fv
#> 3   thetamedian  7.270942e-01           NA        Cox, Gamma      Gamma  fv
#> 4       se2mean  2.033296e-02           NA        Cox, Gamma      Gamma  fv
#> 5     se2median  1.927826e-02           NA        Cox, Gamma      Gamma  fv
#> 6          bias -1.240393e-02 4.548057e-03        Cox, Gamma      Gamma  fv
#> 7         empse  1.420858e-01 3.217611e-03        Cox, Gamma      Gamma  fv
#> 8           mse  2.032156e-02 9.743423e-04        Cox, Gamma      Gamma  fv
#> 9       relprec  1.000000e+00 0.000000e+00        Cox, Gamma      Gamma  fv
#> 10      modelse  1.425937e-01 8.288928e-04        Cox, Gamma      Gamma  fv
#> 11     relerror  3.574233e-01 2.345764e+00        Cox, Gamma      Gamma  fv
#> 12        cover  9.200820e-01 8.679820e-03        Cox, Gamma      Gamma  fv
#> 13      bccover  9.334016e-01 7.980703e-03        Cox, Gamma      Gamma  fv
#> 14        power  1.000000e+00 0.000000e+00        Cox, Gamma      Gamma  fv
#> 15         nsim  1.000000e+03           NA   Cox, Log-Normal      Gamma  fv
#> 16    thetamean  9.799132e-01           NA   Cox, Log-Normal      Gamma  fv
#> 17  thetamedian  9.566223e-01           NA   Cox, Log-Normal      Gamma  fv
#> 18      se2mean  5.998517e-02           NA   Cox, Log-Normal      Gamma  fv
#> 19    se2median  4.832806e-02           NA   Cox, Log-Normal      Gamma  fv
#> 20         bias  2.299132e-01 7.608843e-03   Cox, Log-Normal      Gamma  fv
#> 21        empse  2.406127e-01 5.382956e-03   Cox, Log-Normal      Gamma  fv
#> 22          mse  1.106966e-01 5.471330e-03   Cox, Log-Normal      Gamma  fv
#> 23      relprec  3.487100e-01 6.149322e-03   Cox, Log-Normal      Gamma  fv
#> 24      modelse  2.449187e-01 2.733614e-03   Cox, Log-Normal      Gamma  fv
#> 25     relerror  1.789582e+00 2.544383e+00   Cox, Log-Normal      Gamma  fv
#> 26        cover  9.220000e-01 8.480330e-03   Cox, Log-Normal      Gamma  fv
#> 27      bccover  8.980000e-01 9.570580e-03   Cox, Log-Normal      Gamma  fv
#> 28        power  1.000000e+00 0.000000e+00   Cox, Log-Normal      Gamma  fv
#> 29         nsim  9.710000e+02           NA      RP(P), Gamma      Gamma  fv
#> 30    thetamean  7.320960e-01           NA      RP(P), Gamma      Gamma  fv
#> 31  thetamedian  7.225398e-01           NA      RP(P), Gamma      Gamma  fv
#> 32      se2mean  2.015701e-02           NA      RP(P), Gamma      Gamma  fv
#> 33    se2median  1.912820e-02           NA      RP(P), Gamma      Gamma  fv
#> 34         bias -1.790401e-02 4.449761e-03      RP(P), Gamma      Gamma  fv
#> 35        empse  1.386584e-01 3.148077e-03      RP(P), Gamma      Gamma  fv
#> 36          mse  1.952691e-02 8.825615e-04      RP(P), Gamma      Gamma  fv
#> 37      relprec  1.050048e+00 5.069852e-04      RP(P), Gamma      Gamma  fv
#> 38      modelse  1.419754e-01 7.989133e-04      RP(P), Gamma      Gamma  fv
#> 39     relerror  2.392174e+00 2.394451e+00      RP(P), Gamma      Gamma  fv
#> 40        cover  9.299691e-01 8.189728e-03      RP(P), Gamma      Gamma  fv
#> 41      bccover  9.433574e-01 7.418229e-03      RP(P), Gamma      Gamma  fv
#> 42        power  1.000000e+00 0.000000e+00      RP(P), Gamma      Gamma  fv
#> 43         nsim  1.000000e+03           NA RP(P), Log-Normal      Gamma  fv
#> 44    thetamean  9.847452e-01           NA RP(P), Log-Normal      Gamma  fv
#> 45  thetamedian  9.596519e-01           NA RP(P), Log-Normal      Gamma  fv
#> 46      se2mean  4.981348e-02           NA RP(P), Log-Normal      Gamma  fv
#> 47    se2median  4.423124e-02           NA RP(P), Log-Normal      Gamma  fv
#> 48         bias  2.347452e-01 7.709560e-03 RP(P), Log-Normal      Gamma  fv
#> 49        empse  2.437977e-01 5.454210e-03 RP(P), Log-Normal      Gamma  fv
#> 50          mse  1.144832e-01 5.689660e-03 RP(P), Log-Normal      Gamma  fv
#> 51      relprec  3.396584e-01 5.911277e-03 RP(P), Log-Normal      Gamma  fv
#> 52      modelse  2.231893e-01 1.884023e-03 RP(P), Log-Normal      Gamma  fv
#> 53     relerror -8.453059e+00 2.188540e+00 RP(P), Log-Normal      Gamma  fv
#> 54        cover  9.030000e-01 9.359006e-03 RP(P), Log-Normal      Gamma  fv
#> 55      bccover  8.930000e-01 9.775019e-03 RP(P), Log-Normal      Gamma  fv
#> 56        power  1.000000e+00 0.000000e+00 RP(P), Log-Normal      Gamma  fv
#> 57         nsim  9.570000e+02           NA        Cox, Gamma Log-Normal  fv
#> 58    thetamean  6.435913e-01           NA        Cox, Gamma Log-Normal  fv
#> 59  thetamedian  6.364515e-01           NA        Cox, Gamma Log-Normal  fv
#> 60      se2mean  1.559243e-02           NA        Cox, Gamma Log-Normal  fv
#> 61    se2median  1.490414e-02           NA        Cox, Gamma Log-Normal  fv
#> 62         bias -1.064087e-01 4.266477e-03        Cox, Gamma Log-Normal  fv
#> 63        empse  1.319853e-01 3.018432e-03        Cox, Gamma Log-Normal  fv
#> 64          mse  2.872472e-02 1.042157e-03        Cox, Gamma Log-Normal  fv
#> 65      relprec  1.000000e+00 0.000000e+00        Cox, Gamma Log-Normal  fv
#> 66      modelse  1.248696e-01 7.913984e-04        Cox, Gamma Log-Normal  fv
#> 67     relerror -5.391224e+00 2.244655e+00        Cox, Gamma Log-Normal  fv
#> 68        cover  7.502612e-01 1.399243e-02        Cox, Gamma Log-Normal  fv
#> 69      bccover  9.164054e-01 8.946989e-03        Cox, Gamma Log-Normal  fv
#> 70        power  1.000000e+00 0.000000e+00        Cox, Gamma Log-Normal  fv
#> 71         nsim  1.000000e+03           NA   Cox, Log-Normal Log-Normal  fv
#> 72    thetamean  7.325203e-01           NA   Cox, Log-Normal Log-Normal  fv
#> 73  thetamedian  7.181926e-01           NA   Cox, Log-Normal Log-Normal  fv
#> 74      se2mean  2.302150e-02           NA   Cox, Log-Normal Log-Normal  fv
#> 75    se2median  2.063843e-02           NA   Cox, Log-Normal Log-Normal  fv
#> 76         bias -1.747970e-02 4.912947e-03   Cox, Log-Normal Log-Normal  fv
#> 77        empse  1.553610e-01 3.475716e-03   Cox, Log-Normal Log-Normal  fv
#> 78          mse  2.441845e-02 1.131840e-03   Cox, Log-Normal Log-Normal  fv
#> 79      relprec  7.217166e-01 1.503727e-02   Cox, Log-Normal Log-Normal  fv
#> 80      modelse  1.517284e-01 1.257861e-03   Cox, Log-Normal Log-Normal  fv
#> 81     relerror -2.338199e+00 2.329552e+00   Cox, Log-Normal Log-Normal  fv
#> 82        cover  9.020000e-01 9.401915e-03   Cox, Log-Normal Log-Normal  fv
#> 83      bccover  9.130000e-01 8.912407e-03   Cox, Log-Normal Log-Normal  fv
#> 84        power  1.000000e+00 0.000000e+00   Cox, Log-Normal Log-Normal  fv
#> 85         nsim  9.970000e+02           NA      RP(P), Gamma Log-Normal  fv
#> 86    thetamean  6.434197e-01           NA      RP(P), Gamma Log-Normal  fv
#> 87  thetamedian  6.323696e-01           NA      RP(P), Gamma Log-Normal  fv
#> 88      se2mean  1.582680e-02           NA      RP(P), Gamma Log-Normal  fv
#> 89    se2median  1.491471e-02           NA      RP(P), Gamma Log-Normal  fv
#> 90         bias -1.065803e-01 4.137834e-03      RP(P), Gamma Log-Normal  fv
#> 91        empse  1.306534e-01 2.927359e-03      RP(P), Gamma Log-Normal  fv
#> 92          mse  2.841254e-02 1.002700e-03      RP(P), Gamma Log-Normal  fv
#> 93      relprec  1.020492e+00 4.663717e-04      RP(P), Gamma Log-Normal  fv
#> 94      modelse  1.258046e-01 7.373971e-04      RP(P), Gamma Log-Normal  fv
#> 95     relerror -3.711153e+00 2.229482e+00      RP(P), Gamma Log-Normal  fv
#> 96        cover  7.683049e-01 1.336219e-02      RP(P), Gamma Log-Normal  fv
#> 97      bccover  9.307924e-01 8.038144e-03      RP(P), Gamma Log-Normal  fv
#> 98        power  1.000000e+00 0.000000e+00      RP(P), Gamma Log-Normal  fv
#> 99         nsim  1.000000e+03           NA RP(P), Log-Normal Log-Normal  fv
#> 100   thetamean  7.347699e-01           NA RP(P), Log-Normal Log-Normal  fv
#> 101 thetamedian  7.199359e-01           NA RP(P), Log-Normal Log-Normal  fv
#> 102     se2mean  2.540740e-02           NA RP(P), Log-Normal Log-Normal  fv
#> 103   se2median  2.345964e-02           NA RP(P), Log-Normal Log-Normal  fv
#> 104        bias -1.523014e-02 4.964020e-03 RP(P), Log-Normal Log-Normal  fv
#> 105       empse  1.569761e-01 3.511848e-03 RP(P), Log-Normal Log-Normal  fv
#> 106         mse  2.484881e-02 1.166773e-03 RP(P), Log-Normal Log-Normal  fv
#> 107     relprec  7.069421e-01 1.459120e-02 RP(P), Log-Normal Log-Normal  fv
#> 108     modelse  1.593970e-01 1.067556e-03 RP(P), Log-Normal Log-Normal  fv
#> 109    relerror  1.542208e+00 2.370758e+00 RP(P), Log-Normal Log-Normal  fv
#> 110       cover  9.280000e-01 8.174105e-03 RP(P), Log-Normal Log-Normal  fv
#> 111     bccover  9.360000e-01 7.739767e-03 RP(P), Log-Normal Log-Normal  fv
#> 112       power  1.000000e+00 0.000000e+00 RP(P), Log-Normal Log-Normal  fv
#> 113        nsim  1.000000e+03           NA        Cox, Gamma      Gamma trt
#> 114   thetamean -5.005668e-01           NA        Cox, Gamma      Gamma trt
#> 115 thetamedian -5.010971e-01           NA        Cox, Gamma      Gamma trt
#> 116     se2mean  2.565362e-03           NA        Cox, Gamma      Gamma trt
#> 117   se2median  2.554845e-03           NA        Cox, Gamma      Gamma trt
#> 118        bias -5.668011e-04 1.604912e-03        Cox, Gamma      Gamma trt
#> 119       empse  5.075177e-02 1.135412e-03        Cox, Gamma      Gamma trt
#> 120         mse  2.573488e-03 1.220534e-04        Cox, Gamma      Gamma trt
#> 121     relprec  1.000000e+00 0.000000e+00        Cox, Gamma      Gamma trt
#> 122     modelse  5.064941e-02 4.676077e-05        Cox, Gamma      Gamma trt
#> 123    relerror -2.016972e-01 2.234017e+00        Cox, Gamma      Gamma trt
#> 124       cover  9.500000e-01 6.892024e-03        Cox, Gamma      Gamma trt
#> 125     bccover  9.500000e-01 6.892024e-03        Cox, Gamma      Gamma trt
#> 126       power  1.000000e+00 0.000000e+00        Cox, Gamma      Gamma trt
#> 127        nsim  1.000000e+03           NA   Cox, Log-Normal      Gamma trt
#> 128   thetamean -5.013451e-01           NA   Cox, Log-Normal      Gamma trt
#> 129 thetamedian -5.021194e-01           NA   Cox, Log-Normal      Gamma trt
#> 130     se2mean  2.567182e-03           NA   Cox, Log-Normal      Gamma trt
#> 131   se2median  2.556268e-03           NA   Cox, Log-Normal      Gamma trt
#> 132        bias -1.345050e-03 1.608501e-03   Cox, Log-Normal      Gamma trt
#> 133       empse  5.086525e-02 1.137951e-03   Cox, Log-Normal      Gamma trt
#> 134         mse  2.586496e-03 1.227699e-04   Cox, Log-Normal      Gamma trt
#> 135     relprec  9.955429e-01 1.133244e-03   Cox, Log-Normal      Gamma trt
#> 136     modelse  5.066736e-02 4.683951e-05   Cox, Log-Normal      Gamma trt
#> 137    relerror -3.890478e-01 2.229828e+00   Cox, Log-Normal      Gamma trt
#> 138       cover  9.490000e-01 6.956939e-03   Cox, Log-Normal      Gamma trt
#> 139     bccover  9.500000e-01 6.892024e-03   Cox, Log-Normal      Gamma trt
#> 140       power  1.000000e+00 0.000000e+00   Cox, Log-Normal      Gamma trt
#> 141        nsim  9.710000e+02           NA      RP(P), Gamma      Gamma trt
#> 142   thetamean -5.002832e-01           NA      RP(P), Gamma      Gamma trt
#> 143 thetamedian -5.010410e-01           NA      RP(P), Gamma      Gamma trt
#> 144     se2mean  2.561686e-03           NA      RP(P), Gamma      Gamma trt
#> 145   se2median  2.551366e-03           NA      RP(P), Gamma      Gamma trt
#> 146        bias -2.831604e-04 1.625135e-03      RP(P), Gamma      Gamma trt
#> 147       empse  5.064064e-02 1.149737e-03      RP(P), Gamma      Gamma trt
#> 148         mse  2.561914e-03 1.244542e-04      RP(P), Gamma      Gamma trt
#> 149     relprec  1.004394e+00 8.452122e-04      RP(P), Gamma      Gamma trt
#> 150     modelse  5.061310e-02 4.638305e-05      RP(P), Gamma      Gamma trt
#> 151    relerror -5.438929e-02 2.270412e+00      RP(P), Gamma      Gamma trt
#> 152       cover  9.505664e-01 6.956529e-03      RP(P), Gamma      Gamma trt
#> 153     bccover  9.505664e-01 6.956529e-03      RP(P), Gamma      Gamma trt
#> 154       power  1.000000e+00 0.000000e+00      RP(P), Gamma      Gamma trt
#> 155        nsim  1.000000e+03           NA RP(P), Log-Normal      Gamma trt
#> 156   thetamean -5.015410e-01           NA RP(P), Log-Normal      Gamma trt
#> 157 thetamedian -5.025332e-01           NA RP(P), Log-Normal      Gamma trt
#> 158     se2mean  2.568335e-03           NA RP(P), Log-Normal      Gamma trt
#> 159   se2median  2.558593e-03           NA RP(P), Log-Normal      Gamma trt
#> 160        bias -1.540987e-03 1.609572e-03 RP(P), Log-Normal      Gamma trt
#> 161       empse  5.089912e-02 1.138709e-03 RP(P), Log-Normal      Gamma trt
#> 162         mse  2.590505e-03 1.229553e-04 RP(P), Log-Normal      Gamma trt
#> 163     relprec  9.942184e-01 1.641162e-03 RP(P), Log-Normal      Gamma trt
#> 164     modelse  5.067874e-02 4.679270e-05 RP(P), Log-Normal      Gamma trt
#> 165    relerror -4.329781e-01 2.228840e+00 RP(P), Log-Normal      Gamma trt
#> 166       cover  9.500000e-01 6.892024e-03 RP(P), Log-Normal      Gamma trt
#> 167     bccover  9.490000e-01 6.956939e-03 RP(P), Log-Normal      Gamma trt
#> 168       power  1.000000e+00 0.000000e+00 RP(P), Log-Normal      Gamma trt
#> 169        nsim  1.000000e+03           NA        Cox, Gamma Log-Normal trt
#> 170   thetamean -5.005803e-01           NA        Cox, Gamma Log-Normal trt
#> 171 thetamedian -5.013945e-01           NA        Cox, Gamma Log-Normal trt
#> 172     se2mean  2.254614e-03           NA        Cox, Gamma Log-Normal trt
#> 173   se2median  2.244556e-03           NA        Cox, Gamma Log-Normal trt
#> 174        bias -5.802773e-04 1.497783e-03        Cox, Gamma Log-Normal trt
#> 175       empse  4.736405e-02 1.059622e-03        Cox, Gamma Log-Normal trt
#> 176         mse  2.241447e-03 1.010505e-04        Cox, Gamma Log-Normal trt
#> 177     relprec  1.000000e+00 1.333467e-09        Cox, Gamma Log-Normal trt
#> 178     modelse  4.748277e-02 3.257581e-05        Cox, Gamma Log-Normal trt
#> 179    relerror  2.506527e-01 2.243288e+00        Cox, Gamma Log-Normal trt
#> 180       cover  9.410000e-01 7.451107e-03        Cox, Gamma Log-Normal trt
#> 181     bccover  9.420000e-01 7.391617e-03        Cox, Gamma Log-Normal trt
#> 182       power  1.000000e+00 0.000000e+00        Cox, Gamma Log-Normal trt
#> 183        nsim  1.000000e+03           NA   Cox, Log-Normal Log-Normal trt
#> 184   thetamean -5.014181e-01           NA   Cox, Log-Normal Log-Normal trt
#> 185 thetamedian -5.021300e-01           NA   Cox, Log-Normal Log-Normal trt
#> 186     se2mean  2.254699e-03           NA   Cox, Log-Normal Log-Normal trt
#> 187   se2median  2.244851e-03           NA   Cox, Log-Normal Log-Normal trt
#> 188        bias -1.418120e-03 1.498845e-03   Cox, Log-Normal Log-Normal trt
#> 189       empse  4.739765e-02 1.060374e-03   Cox, Log-Normal Log-Normal trt
#> 190         mse  2.246302e-03 1.015180e-04   Cox, Log-Normal Log-Normal trt
#> 191     relprec  9.985828e-01 1.367338e-03   Cox, Log-Normal Log-Normal trt
#> 192     modelse  4.748367e-02 3.266741e-05   Cox, Log-Normal Log-Normal trt
#> 193    relerror  1.814780e-01 2.241746e+00   Cox, Log-Normal Log-Normal trt
#> 194       cover  9.420000e-01 7.391617e-03   Cox, Log-Normal Log-Normal trt
#> 195     bccover  9.400000e-01 7.509993e-03   Cox, Log-Normal Log-Normal trt
#> 196       power  1.000000e+00 0.000000e+00   Cox, Log-Normal Log-Normal trt
#> 197        nsim  9.970000e+02           NA      RP(P), Gamma Log-Normal trt
#> 198   thetamean -5.006488e-01           NA      RP(P), Gamma Log-Normal trt
#> 199 thetamedian -5.014055e-01           NA      RP(P), Gamma Log-Normal trt
#> 200     se2mean  2.256201e-03           NA      RP(P), Gamma Log-Normal trt
#> 201   se2median  2.245797e-03           NA      RP(P), Gamma Log-Normal trt
#> 202        bias -6.488363e-04 1.499347e-03      RP(P), Gamma Log-Normal trt
#> 203       empse  4.734234e-02 1.060730e-03      RP(P), Gamma Log-Normal trt
#> 204         mse  2.239470e-03 1.013281e-04      RP(P), Gamma Log-Normal trt
#> 205     relprec  1.000918e+00 8.532518e-04      RP(P), Gamma Log-Normal trt
#> 206     modelse  4.749949e-02 3.256088e-05      RP(P), Gamma Log-Normal trt
#> 207    relerror  3.319367e-01 2.248479e+00      RP(P), Gamma Log-Normal trt
#> 208       cover  9.428285e-01 7.352899e-03      RP(P), Gamma Log-Normal trt
#> 209     bccover  9.428285e-01 7.352899e-03      RP(P), Gamma Log-Normal trt
#> 210       power  1.000000e+00 0.000000e+00      RP(P), Gamma Log-Normal trt
#> 211        nsim  1.000000e+03           NA RP(P), Log-Normal Log-Normal trt
#> 212   thetamean -5.015887e-01           NA RP(P), Log-Normal Log-Normal trt
#> 213 thetamedian -5.022269e-01           NA RP(P), Log-Normal Log-Normal trt
#> 214     se2mean  2.257481e-03           NA RP(P), Log-Normal Log-Normal trt
#> 215   se2median  2.247168e-03           NA RP(P), Log-Normal Log-Normal trt
#> 216        bias -1.588678e-03 1.499342e-03 RP(P), Log-Normal Log-Normal trt
#> 217       empse  4.741334e-02 1.060725e-03 RP(P), Log-Normal Log-Normal trt
#> 218         mse  2.248301e-03 1.015869e-04 RP(P), Log-Normal Log-Normal trt
#> 219     relprec  9.979220e-01 1.588609e-03 RP(P), Log-Normal Log-Normal trt
#> 220     modelse  4.751296e-02 3.262005e-05 RP(P), Log-Normal Log-Normal trt
#> 221    relerror  2.100926e-01 2.242382e+00 RP(P), Log-Normal Log-Normal trt
#> 222       cover  9.430000e-01 7.331507e-03 RP(P), Log-Normal Log-Normal trt
#> 223     bccover  9.410000e-01 7.451107e-03 RP(P), Log-Normal Log-Normal trt
#> 224       power  1.000000e+00 0.000000e+00 RP(P), Log-Normal Log-Normal trt

get_data(ms, description = TRUE)

#>            stat           est         mcse             model    fv_dist par
#> 1       bccover  9.433574e-01 7.418229e-03      RP(P), Gamma      Gamma  fv
#> 2       bccover  9.164054e-01 8.946989e-03        Cox, Gamma Log-Normal  fv
#> 3       bccover  9.130000e-01 8.912407e-03   Cox, Log-Normal Log-Normal  fv
#> 4       bccover  9.500000e-01 6.892024e-03        Cox, Gamma      Gamma trt
#> 5       bccover  9.505664e-01 6.956529e-03      RP(P), Gamma      Gamma trt
#> 6       bccover  9.490000e-01 6.956939e-03 RP(P), Log-Normal      Gamma trt
#> 7       bccover  9.360000e-01 7.739767e-03 RP(P), Log-Normal Log-Normal  fv
#> 8       bccover  9.428285e-01 7.352899e-03      RP(P), Gamma Log-Normal trt
#> 9       bccover  9.334016e-01 7.980703e-03        Cox, Gamma      Gamma  fv
#> 10      bccover  8.930000e-01 9.775019e-03 RP(P), Log-Normal      Gamma  fv
#> 11      bccover  9.400000e-01 7.509993e-03   Cox, Log-Normal Log-Normal trt
#> 12      bccover  8.980000e-01 9.570580e-03   Cox, Log-Normal      Gamma  fv
#> 13      bccover  9.500000e-01 6.892024e-03   Cox, Log-Normal      Gamma trt
#> 14      bccover  9.307924e-01 8.038144e-03      RP(P), Gamma Log-Normal  fv
#> 15      bccover  9.410000e-01 7.451107e-03 RP(P), Log-Normal Log-Normal trt
#> 16      bccover  9.420000e-01 7.391617e-03        Cox, Gamma Log-Normal trt
#> 17         bias -6.488363e-04 1.499347e-03      RP(P), Gamma Log-Normal trt
#> 18         bias -1.418120e-03 1.498845e-03   Cox, Log-Normal Log-Normal trt
#> 19         bias -1.540987e-03 1.609572e-03 RP(P), Log-Normal      Gamma trt
#> 20         bias -1.588678e-03 1.499342e-03 RP(P), Log-Normal Log-Normal trt
#> 21         bias  2.347452e-01 7.709560e-03 RP(P), Log-Normal      Gamma  fv
#> 22         bias -5.802773e-04 1.497783e-03        Cox, Gamma Log-Normal trt
#> 23         bias -2.831604e-04 1.625135e-03      RP(P), Gamma      Gamma trt
#> 24         bias -1.345050e-03 1.608501e-03   Cox, Log-Normal      Gamma trt
#> 25         bias -5.668011e-04 1.604912e-03        Cox, Gamma      Gamma trt
#> 26         bias -1.790401e-02 4.449761e-03      RP(P), Gamma      Gamma  fv
#> 27         bias -1.240393e-02 4.548057e-03        Cox, Gamma      Gamma  fv
#> 28         bias -1.747970e-02 4.912947e-03   Cox, Log-Normal Log-Normal  fv
#> 29         bias  2.299132e-01 7.608843e-03   Cox, Log-Normal      Gamma  fv
#> 30         bias -1.065803e-01 4.137834e-03      RP(P), Gamma Log-Normal  fv
#> 31         bias -1.523014e-02 4.964020e-03 RP(P), Log-Normal Log-Normal  fv
#> 32         bias -1.064087e-01 4.266477e-03        Cox, Gamma Log-Normal  fv
#> 33        cover  9.280000e-01 8.174105e-03 RP(P), Log-Normal Log-Normal  fv
#> 34        cover  7.502612e-01 1.399243e-02        Cox, Gamma Log-Normal  fv
#> 35        cover  9.420000e-01 7.391617e-03   Cox, Log-Normal Log-Normal trt
#> 36        cover  7.683049e-01 1.336219e-02      RP(P), Gamma Log-Normal  fv
#> 37        cover  9.505664e-01 6.956529e-03      RP(P), Gamma      Gamma trt
#> 38        cover  9.030000e-01 9.359006e-03 RP(P), Log-Normal      Gamma  fv
#> 39        cover  9.020000e-01 9.401915e-03   Cox, Log-Normal Log-Normal  fv
#> 40        cover  9.410000e-01 7.451107e-03        Cox, Gamma Log-Normal trt
#> 41        cover  9.500000e-01 6.892024e-03        Cox, Gamma      Gamma trt
#> 42        cover  9.490000e-01 6.956939e-03   Cox, Log-Normal      Gamma trt
#> 43        cover  9.500000e-01 6.892024e-03 RP(P), Log-Normal      Gamma trt
#> 44        cover  9.200820e-01 8.679820e-03        Cox, Gamma      Gamma  fv
#> 45        cover  9.428285e-01 7.352899e-03      RP(P), Gamma Log-Normal trt
#> 46        cover  9.430000e-01 7.331507e-03 RP(P), Log-Normal Log-Normal trt
#> 47        cover  9.299691e-01 8.189728e-03      RP(P), Gamma      Gamma  fv
#> 48        cover  9.220000e-01 8.480330e-03   Cox, Log-Normal      Gamma  fv
#> 49        empse  4.734234e-02 1.060730e-03      RP(P), Gamma Log-Normal trt
#> 50        empse  5.089912e-02 1.138709e-03 RP(P), Log-Normal      Gamma trt
#> 51        empse  4.736405e-02 1.059622e-03        Cox, Gamma Log-Normal trt
#> 52        empse  5.064064e-02 1.149737e-03      RP(P), Gamma      Gamma trt
#> 53        empse  4.739765e-02 1.060374e-03   Cox, Log-Normal Log-Normal trt
#> 54        empse  5.075177e-02 1.135412e-03        Cox, Gamma      Gamma trt
#> 55        empse  1.420858e-01 3.217611e-03        Cox, Gamma      Gamma  fv
#> 56        empse  4.741334e-02 1.060725e-03 RP(P), Log-Normal Log-Normal trt
#> 57        empse  5.086525e-02 1.137951e-03   Cox, Log-Normal      Gamma trt
#> 58        empse  1.569761e-01 3.511848e-03 RP(P), Log-Normal Log-Normal  fv
#> 59        empse  1.306534e-01 2.927359e-03      RP(P), Gamma Log-Normal  fv
#> 60        empse  2.406127e-01 5.382956e-03   Cox, Log-Normal      Gamma  fv
#> 61        empse  2.437977e-01 5.454210e-03 RP(P), Log-Normal      Gamma  fv
#> 62        empse  1.319853e-01 3.018432e-03        Cox, Gamma Log-Normal  fv
#> 63        empse  1.553610e-01 3.475716e-03   Cox, Log-Normal Log-Normal  fv
#> 64        empse  1.386584e-01 3.148077e-03      RP(P), Gamma      Gamma  fv
#> 65      modelse  5.061310e-02 4.638305e-05      RP(P), Gamma      Gamma trt
#> 66      modelse  1.593970e-01 1.067556e-03 RP(P), Log-Normal Log-Normal  fv
#> 67      modelse  5.064941e-02 4.676077e-05        Cox, Gamma      Gamma trt
#> 68      modelse  5.066736e-02 4.683951e-05   Cox, Log-Normal      Gamma trt
#> 69      modelse  4.748367e-02 3.266741e-05   Cox, Log-Normal Log-Normal trt
#> 70      modelse  1.258046e-01 7.373971e-04      RP(P), Gamma Log-Normal  fv
#> 71      modelse  4.749949e-02 3.256088e-05      RP(P), Gamma Log-Normal trt
#> 72      modelse  4.751296e-02 3.262005e-05 RP(P), Log-Normal Log-Normal trt
#> 73      modelse  5.067874e-02 4.679270e-05 RP(P), Log-Normal      Gamma trt
#> 74      modelse  4.748277e-02 3.257581e-05        Cox, Gamma Log-Normal trt
#> 75      modelse  1.517284e-01 1.257861e-03   Cox, Log-Normal Log-Normal  fv
#> 76      modelse  2.231893e-01 1.884023e-03 RP(P), Log-Normal      Gamma  fv
#> 77      modelse  1.419754e-01 7.989133e-04      RP(P), Gamma      Gamma  fv
#> 78      modelse  1.248696e-01 7.913984e-04        Cox, Gamma Log-Normal  fv
#> 79      modelse  2.449187e-01 2.733614e-03   Cox, Log-Normal      Gamma  fv
#> 80      modelse  1.425937e-01 8.288928e-04        Cox, Gamma      Gamma  fv
#> 81          mse  2.246302e-03 1.015180e-04   Cox, Log-Normal Log-Normal trt
#> 82          mse  2.561914e-03 1.244542e-04      RP(P), Gamma      Gamma trt
#> 83          mse  2.590505e-03 1.229553e-04 RP(P), Log-Normal      Gamma trt
#> 84          mse  2.241447e-03 1.010505e-04        Cox, Gamma Log-Normal trt
#> 85          mse  2.239470e-03 1.013281e-04      RP(P), Gamma Log-Normal trt
#> 86          mse  2.586496e-03 1.227699e-04   Cox, Log-Normal      Gamma trt
#> 87          mse  2.573488e-03 1.220534e-04        Cox, Gamma      Gamma trt
#> 88          mse  2.841254e-02 1.002700e-03      RP(P), Gamma Log-Normal  fv
#> 89          mse  2.032156e-02 9.743423e-04        Cox, Gamma      Gamma  fv
#> 90          mse  2.248301e-03 1.015869e-04 RP(P), Log-Normal Log-Normal trt
#> 91          mse  2.484881e-02 1.166773e-03 RP(P), Log-Normal Log-Normal  fv
#> 92          mse  1.952691e-02 8.825615e-04      RP(P), Gamma      Gamma  fv
#> 93          mse  2.441845e-02 1.131840e-03   Cox, Log-Normal Log-Normal  fv
#> 94          mse  1.144832e-01 5.689660e-03 RP(P), Log-Normal      Gamma  fv
#> 95          mse  2.872472e-02 1.042157e-03        Cox, Gamma Log-Normal  fv
#> 96          mse  1.106966e-01 5.471330e-03   Cox, Log-Normal      Gamma  fv
#> 97         nsim  9.760000e+02           NA        Cox, Gamma      Gamma  fv
#> 98         nsim  1.000000e+03           NA   Cox, Log-Normal      Gamma  fv
#> 99         nsim  1.000000e+03           NA RP(P), Log-Normal      Gamma  fv
#> 100        nsim  9.570000e+02           NA        Cox, Gamma Log-Normal  fv
#> 101        nsim  9.970000e+02           NA      RP(P), Gamma Log-Normal  fv
#> 102        nsim  1.000000e+03           NA RP(P), Log-Normal Log-Normal  fv
#> 103        nsim  1.000000e+03           NA   Cox, Log-Normal      Gamma trt
#> 104        nsim  9.710000e+02           NA      RP(P), Gamma      Gamma trt
#> 105        nsim  1.000000e+03           NA        Cox, Gamma Log-Normal trt
#> 106        nsim  1.000000e+03           NA   Cox, Log-Normal Log-Normal trt
#> 107        nsim  1.000000e+03           NA RP(P), Log-Normal Log-Normal trt
#> 108        nsim  9.710000e+02           NA      RP(P), Gamma      Gamma  fv
#> 109        nsim  1.000000e+03           NA   Cox, Log-Normal Log-Normal  fv
#> 110        nsim  1.000000e+03           NA        Cox, Gamma      Gamma trt
#> 111        nsim  1.000000e+03           NA RP(P), Log-Normal      Gamma trt
#> 112        nsim  9.970000e+02           NA      RP(P), Gamma Log-Normal trt
#> 113       power  1.000000e+00 0.000000e+00   Cox, Log-Normal      Gamma  fv
#> 114       power  1.000000e+00 0.000000e+00        Cox, Gamma Log-Normal  fv
#> 115       power  1.000000e+00 0.000000e+00 RP(P), Log-Normal Log-Normal  fv
#> 116       power  1.000000e+00 0.000000e+00      RP(P), Gamma      Gamma trt
#> 117       power  1.000000e+00 0.000000e+00   Cox, Log-Normal Log-Normal trt
#> 118       power  1.000000e+00 0.000000e+00        Cox, Gamma      Gamma  fv
#> 119       power  1.000000e+00 0.000000e+00      RP(P), Gamma      Gamma  fv
#> 120       power  1.000000e+00 0.000000e+00 RP(P), Log-Normal      Gamma  fv
#> 121       power  1.000000e+00 0.000000e+00   Cox, Log-Normal Log-Normal  fv
#> 122       power  1.000000e+00 0.000000e+00      RP(P), Gamma Log-Normal  fv
#> 123       power  1.000000e+00 0.000000e+00        Cox, Gamma      Gamma trt
#> 124       power  1.000000e+00 0.000000e+00   Cox, Log-Normal      Gamma trt
#> 125       power  1.000000e+00 0.000000e+00 RP(P), Log-Normal      Gamma trt
#> 126       power  1.000000e+00 0.000000e+00        Cox, Gamma Log-Normal trt
#> 127       power  1.000000e+00 0.000000e+00      RP(P), Gamma Log-Normal trt
#> 128       power  1.000000e+00 0.000000e+00 RP(P), Log-Normal Log-Normal trt
#> 129    relerror -2.016972e-01 2.234017e+00        Cox, Gamma      Gamma trt
#> 130    relerror -2.338199e+00 2.329552e+00   Cox, Log-Normal Log-Normal  fv
#> 131    relerror  3.319367e-01 2.248479e+00      RP(P), Gamma Log-Normal trt
#> 132    relerror -4.329781e-01 2.228840e+00 RP(P), Log-Normal      Gamma trt
#> 133    relerror  1.542208e+00 2.370758e+00 RP(P), Log-Normal Log-Normal  fv
#> 134    relerror -3.711153e+00 2.229482e+00      RP(P), Gamma Log-Normal  fv
#> 135    relerror -5.438929e-02 2.270412e+00      RP(P), Gamma      Gamma trt
#> 136    relerror -5.391224e+00 2.244655e+00        Cox, Gamma Log-Normal  fv
#> 137    relerror  1.814780e-01 2.241746e+00   Cox, Log-Normal Log-Normal trt
#> 138    relerror  2.506527e-01 2.243288e+00        Cox, Gamma Log-Normal trt
#> 139    relerror  3.574233e-01 2.345764e+00        Cox, Gamma      Gamma  fv
#> 140    relerror -3.890478e-01 2.229828e+00   Cox, Log-Normal      Gamma trt
#> 141    relerror -8.453059e+00 2.188540e+00 RP(P), Log-Normal      Gamma  fv
#> 142    relerror  2.392174e+00 2.394451e+00      RP(P), Gamma      Gamma  fv
#> 143    relerror  2.100926e-01 2.242382e+00 RP(P), Log-Normal Log-Normal trt
#> 144    relerror  1.789582e+00 2.544383e+00   Cox, Log-Normal      Gamma  fv
#> 145     relprec  9.942184e-01 1.641162e-03 RP(P), Log-Normal      Gamma trt
#> 146     relprec  1.000000e+00 0.000000e+00        Cox, Gamma      Gamma trt
#> 147     relprec  9.955429e-01 1.133244e-03   Cox, Log-Normal      Gamma trt
#> 148     relprec  1.000918e+00 8.532518e-04      RP(P), Gamma Log-Normal trt
#> 149     relprec  1.004394e+00 8.452122e-04      RP(P), Gamma      Gamma trt
#> 150     relprec  9.979220e-01 1.588609e-03 RP(P), Log-Normal Log-Normal trt
#> 151     relprec  9.985828e-01 1.367338e-03   Cox, Log-Normal Log-Normal trt
#> 152     relprec  7.069421e-01 1.459120e-02 RP(P), Log-Normal Log-Normal  fv
#> 153     relprec  1.020492e+00 4.663717e-04      RP(P), Gamma Log-Normal  fv
#> 154     relprec  7.217166e-01 1.503727e-02   Cox, Log-Normal Log-Normal  fv
#> 155     relprec  3.396584e-01 5.911277e-03 RP(P), Log-Normal      Gamma  fv
#> 156     relprec  1.000000e+00 1.333467e-09        Cox, Gamma Log-Normal trt
#> 157     relprec  1.000000e+00 0.000000e+00        Cox, Gamma      Gamma  fv
#> 158     relprec  1.000000e+00 0.000000e+00        Cox, Gamma Log-Normal  fv
#> 159     relprec  1.050048e+00 5.069852e-04      RP(P), Gamma      Gamma  fv
#> 160     relprec  3.487100e-01 6.149322e-03   Cox, Log-Normal      Gamma  fv
#> 161     se2mean  2.256201e-03           NA      RP(P), Gamma Log-Normal trt
#> 162     se2mean  2.033296e-02           NA        Cox, Gamma      Gamma  fv
#> 163     se2mean  2.254699e-03           NA   Cox, Log-Normal Log-Normal trt
#> 164     se2mean  2.254614e-03           NA        Cox, Gamma Log-Normal trt
#> 165     se2mean  1.582680e-02           NA      RP(P), Gamma Log-Normal  fv
#> 166     se2mean  2.257481e-03           NA RP(P), Log-Normal Log-Normal trt
#> 167     se2mean  4.981348e-02           NA RP(P), Log-Normal      Gamma  fv
#> 168     se2mean  2.015701e-02           NA      RP(P), Gamma      Gamma  fv
#> 169     se2mean  2.568335e-03           NA RP(P), Log-Normal      Gamma trt
#> 170     se2mean  2.302150e-02           NA   Cox, Log-Normal Log-Normal  fv
#> 171     se2mean  2.567182e-03           NA   Cox, Log-Normal      Gamma trt
#> 172     se2mean  2.565362e-03           NA        Cox, Gamma      Gamma trt
#> 173     se2mean  1.559243e-02           NA        Cox, Gamma Log-Normal  fv
#> 174     se2mean  5.998517e-02           NA   Cox, Log-Normal      Gamma  fv
#> 175     se2mean  2.561686e-03           NA      RP(P), Gamma      Gamma trt
#> 176     se2mean  2.540740e-02           NA RP(P), Log-Normal Log-Normal  fv
#> 177   se2median  2.247168e-03           NA RP(P), Log-Normal Log-Normal trt
#> 178   se2median  2.245797e-03           NA      RP(P), Gamma Log-Normal trt
#> 179   se2median  2.558593e-03           NA RP(P), Log-Normal      Gamma trt
#> 180   se2median  2.244851e-03           NA   Cox, Log-Normal Log-Normal trt
#> 181   se2median  2.244556e-03           NA        Cox, Gamma Log-Normal trt
#> 182   se2median  2.551366e-03           NA      RP(P), Gamma      Gamma trt
#> 183   se2median  4.423124e-02           NA RP(P), Log-Normal      Gamma  fv
#> 184   se2median  1.490414e-02           NA        Cox, Gamma Log-Normal  fv
#> 185   se2median  1.927826e-02           NA        Cox, Gamma      Gamma  fv
#> 186   se2median  4.832806e-02           NA   Cox, Log-Normal      Gamma  fv
#> 187   se2median  2.556268e-03           NA   Cox, Log-Normal      Gamma trt
#> 188   se2median  1.912820e-02           NA      RP(P), Gamma      Gamma  fv
#> 189   se2median  1.491471e-02           NA      RP(P), Gamma Log-Normal  fv
#> 190   se2median  2.345964e-02           NA RP(P), Log-Normal Log-Normal  fv
#> 191   se2median  2.554845e-03           NA        Cox, Gamma      Gamma trt
#> 192   se2median  2.063843e-02           NA   Cox, Log-Normal Log-Normal  fv
#> 193   thetamean  9.847452e-01           NA RP(P), Log-Normal      Gamma  fv
#> 194   thetamean  7.375961e-01           NA        Cox, Gamma      Gamma  fv
#> 195   thetamean -5.013451e-01           NA   Cox, Log-Normal      Gamma trt
#> 196   thetamean  6.434197e-01           NA      RP(P), Gamma Log-Normal  fv
#> 197   thetamean -5.006488e-01           NA      RP(P), Gamma Log-Normal trt
#> 198   thetamean  7.320960e-01           NA      RP(P), Gamma      Gamma  fv
#> 199   thetamean -5.005803e-01           NA        Cox, Gamma Log-Normal trt
#> 200   thetamean -5.015887e-01           NA RP(P), Log-Normal Log-Normal trt
#> 201   thetamean  9.799132e-01           NA   Cox, Log-Normal      Gamma  fv
#> 202   thetamean -5.005668e-01           NA        Cox, Gamma      Gamma trt
#> 203   thetamean  6.435913e-01           NA        Cox, Gamma Log-Normal  fv
#> 204   thetamean  7.325203e-01           NA   Cox, Log-Normal Log-Normal  fv
#> 205   thetamean  7.347699e-01           NA RP(P), Log-Normal Log-Normal  fv
#> 206   thetamean -5.002832e-01           NA      RP(P), Gamma      Gamma trt
#> 207   thetamean -5.015410e-01           NA RP(P), Log-Normal      Gamma trt
#> 208   thetamean -5.014181e-01           NA   Cox, Log-Normal Log-Normal trt
#> 209 thetamedian -5.014055e-01           NA      RP(P), Gamma Log-Normal trt
#> 210 thetamedian -5.021300e-01           NA   Cox, Log-Normal Log-Normal trt
#> 211 thetamedian  7.270942e-01           NA        Cox, Gamma      Gamma  fv
#> 212 thetamedian  9.566223e-01           NA   Cox, Log-Normal      Gamma  fv
#> 213 thetamedian -5.022269e-01           NA RP(P), Log-Normal Log-Normal trt
#> 214 thetamedian  6.364515e-01           NA        Cox, Gamma Log-Normal  fv
#> 215 thetamedian  6.323696e-01           NA      RP(P), Gamma Log-Normal  fv
#> 216 thetamedian  7.199359e-01           NA RP(P), Log-Normal Log-Normal  fv
#> 217 thetamedian  9.596519e-01           NA RP(P), Log-Normal      Gamma  fv
#> 218 thetamedian -5.010410e-01           NA      RP(P), Gamma      Gamma trt
#> 219 thetamedian -5.013945e-01           NA        Cox, Gamma Log-Normal trt
#> 220 thetamedian  7.181926e-01           NA   Cox, Log-Normal Log-Normal  fv
#> 221 thetamedian -5.021194e-01           NA   Cox, Log-Normal      Gamma trt
#> 222 thetamedian  7.225398e-01           NA      RP(P), Gamma      Gamma  fv
#> 223 thetamedian -5.025332e-01           NA RP(P), Log-Normal      Gamma trt
#> 224 thetamedian -5.010971e-01           NA        Cox, Gamma      Gamma trt
#>                                           description
#> 1           Bias corrected coverage of nominal 95% CI
#> 2           Bias corrected coverage of nominal 95% CI
#> 3           Bias corrected coverage of nominal 95% CI
#> 4           Bias corrected coverage of nominal 95% CI
#> 5           Bias corrected coverage of nominal 95% CI
#> 6           Bias corrected coverage of nominal 95% CI
#> 7           Bias corrected coverage of nominal 95% CI
#> 8           Bias corrected coverage of nominal 95% CI
#> 9           Bias corrected coverage of nominal 95% CI
#> 10          Bias corrected coverage of nominal 95% CI
#> 11          Bias corrected coverage of nominal 95% CI
#> 12          Bias corrected coverage of nominal 95% CI
#> 13          Bias corrected coverage of nominal 95% CI
#> 14          Bias corrected coverage of nominal 95% CI
#> 15          Bias corrected coverage of nominal 95% CI
#> 16          Bias corrected coverage of nominal 95% CI
#> 17                             Bias in point estimate
#> 18                             Bias in point estimate
#> 19                             Bias in point estimate
#> 20                             Bias in point estimate
#> 21                             Bias in point estimate
#> 22                             Bias in point estimate
#> 23                             Bias in point estimate
#> 24                             Bias in point estimate
#> 25                             Bias in point estimate
#> 26                             Bias in point estimate
#> 27                             Bias in point estimate
#> 28                             Bias in point estimate
#> 29                             Bias in point estimate
#> 30                             Bias in point estimate
#> 31                             Bias in point estimate
#> 32                             Bias in point estimate
#> 33                         Coverage of nominal 95% CI
#> 34                         Coverage of nominal 95% CI
#> 35                         Coverage of nominal 95% CI
#> 36                         Coverage of nominal 95% CI
#> 37                         Coverage of nominal 95% CI
#> 38                         Coverage of nominal 95% CI
#> 39                         Coverage of nominal 95% CI
#> 40                         Coverage of nominal 95% CI
#> 41                         Coverage of nominal 95% CI
#> 42                         Coverage of nominal 95% CI
#> 43                         Coverage of nominal 95% CI
#> 44                         Coverage of nominal 95% CI
#> 45                         Coverage of nominal 95% CI
#> 46                         Coverage of nominal 95% CI
#> 47                         Coverage of nominal 95% CI
#> 48                         Coverage of nominal 95% CI
#> 49                           Empirical standard error
#> 50                           Empirical standard error
#> 51                           Empirical standard error
#> 52                           Empirical standard error
#> 53                           Empirical standard error
#> 54                           Empirical standard error
#> 55                           Empirical standard error
#> 56                           Empirical standard error
#> 57                           Empirical standard error
#> 58                           Empirical standard error
#> 59                           Empirical standard error
#> 60                           Empirical standard error
#> 61                           Empirical standard error
#> 62                           Empirical standard error
#> 63                           Empirical standard error
#> 64                           Empirical standard error
#> 65                         Model-based standard error
#> 66                         Model-based standard error
#> 67                         Model-based standard error
#> 68                         Model-based standard error
#> 69                         Model-based standard error
#> 70                         Model-based standard error
#> 71                         Model-based standard error
#> 72                         Model-based standard error
#> 73                         Model-based standard error
#> 74                         Model-based standard error
#> 75                         Model-based standard error
#> 76                         Model-based standard error
#> 77                         Model-based standard error
#> 78                         Model-based standard error
#> 79                         Model-based standard error
#> 80                         Model-based standard error
#> 81                                 Mean squared error
#> 82                                 Mean squared error
#> 83                                 Mean squared error
#> 84                                 Mean squared error
#> 85                                 Mean squared error
#> 86                                 Mean squared error
#> 87                                 Mean squared error
#> 88                                 Mean squared error
#> 89                                 Mean squared error
#> 90                                 Mean squared error
#> 91                                 Mean squared error
#> 92                                 Mean squared error
#> 93                                 Mean squared error
#> 94                                 Mean squared error
#> 95                                 Mean squared error
#> 96                                 Mean squared error
#> 97         Simulations with non-missing estimates/SEs
#> 98         Simulations with non-missing estimates/SEs
#> 99         Simulations with non-missing estimates/SEs
#> 100        Simulations with non-missing estimates/SEs
#> 101        Simulations with non-missing estimates/SEs
#> 102        Simulations with non-missing estimates/SEs
#> 103        Simulations with non-missing estimates/SEs
#> 104        Simulations with non-missing estimates/SEs
#> 105        Simulations with non-missing estimates/SEs
#> 106        Simulations with non-missing estimates/SEs
#> 107        Simulations with non-missing estimates/SEs
#> 108        Simulations with non-missing estimates/SEs
#> 109        Simulations with non-missing estimates/SEs
#> 110        Simulations with non-missing estimates/SEs
#> 111        Simulations with non-missing estimates/SEs
#> 112        Simulations with non-missing estimates/SEs
#> 113                            Power of 5% level test
#> 114                            Power of 5% level test
#> 115                            Power of 5% level test
#> 116                            Power of 5% level test
#> 117                            Power of 5% level test
#> 118                            Power of 5% level test
#> 119                            Power of 5% level test
#> 120                            Power of 5% level test
#> 121                            Power of 5% level test
#> 122                            Power of 5% level test
#> 123                            Power of 5% level test
#> 124                            Power of 5% level test
#> 125                            Power of 5% level test
#> 126                            Power of 5% level test
#> 127                            Power of 5% level test
#> 128                            Power of 5% level test
#> 129                Relative % error in standard error
#> 130                Relative % error in standard error
#> 131                Relative % error in standard error
#> 132                Relative % error in standard error
#> 133                Relative % error in standard error
#> 134                Relative % error in standard error
#> 135                Relative % error in standard error
#> 136                Relative % error in standard error
#> 137                Relative % error in standard error
#> 138                Relative % error in standard error
#> 139                Relative % error in standard error
#> 140                Relative % error in standard error
#> 141                Relative % error in standard error
#> 142                Relative % error in standard error
#> 143                Relative % error in standard error
#> 144                Relative % error in standard error
#> 145 % gain in precision relative to method Cox, Gamma
#> 146 % gain in precision relative to method Cox, Gamma
#> 147 % gain in precision relative to method Cox, Gamma
#> 148 % gain in precision relative to method Cox, Gamma
#> 149 % gain in precision relative to method Cox, Gamma
#> 150 % gain in precision relative to method Cox, Gamma
#> 151 % gain in precision relative to method Cox, Gamma
#> 152 % gain in precision relative to method Cox, Gamma
#> 153 % gain in precision relative to method Cox, Gamma
#> 154 % gain in precision relative to method Cox, Gamma
#> 155 % gain in precision relative to method Cox, Gamma
#> 156 % gain in precision relative to method Cox, Gamma
#> 157 % gain in precision relative to method Cox, Gamma
#> 158 % gain in precision relative to method Cox, Gamma
#> 159 % gain in precision relative to method Cox, Gamma
#> 160 % gain in precision relative to method Cox, Gamma
#> 161                            Average standard error
#> 162                            Average standard error
#> 163                            Average standard error
#> 164                            Average standard error
#> 165                            Average standard error
#> 166                            Average standard error
#> 167                            Average standard error
#> 168                            Average standard error
#> 169                            Average standard error
#> 170                            Average standard error
#> 171                            Average standard error
#> 172                            Average standard error
#> 173                            Average standard error
#> 174                            Average standard error
#> 175                            Average standard error
#> 176                            Average standard error
#> 177                             Median standard error
#> 178                             Median standard error
#> 179                             Median standard error
#> 180                             Median standard error
#> 181                             Median standard error
#> 182                             Median standard error
#> 183                             Median standard error
#> 184                             Median standard error
#> 185                             Median standard error
#> 186                             Median standard error
#> 187                             Median standard error
#> 188                             Median standard error
#> 189                             Median standard error
#> 190                             Median standard error
#> 191                             Median standard error
#> 192                             Median standard error
#> 193                            Average point estimate
#> 194                            Average point estimate
#> 195                            Average point estimate
#> 196                            Average point estimate
#> 197                            Average point estimate
#> 198                            Average point estimate
#> 199                            Average point estimate
#> 200                            Average point estimate
#> 201                            Average point estimate
#> 202                            Average point estimate
#> 203                            Average point estimate
#> 204                            Average point estimate
#> 205                            Average point estimate
#> 206                            Average point estimate
#> 207                            Average point estimate
#> 208                            Average point estimate
#> 209                             Median point estimate
#> 210                             Median point estimate
#> 211                             Median point estimate
#> 212                             Median point estimate
#> 213                             Median point estimate
#> 214                             Median point estimate
#> 215                             Median point estimate
#> 216                             Median point estimate
#> 217                             Median point estimate
#> 218                             Median point estimate
#> 219                             Median point estimate
#> 220                             Median point estimate
#> 221                             Median point estimate
#> 222                             Median point estimate
#> 223                             Median point estimate
#> 224                             Median point estimate