get_data.summary.simsum

Extract data from a summary.simsum object

# S3 method for summary.simsum
get_data(x, sstat = "all",
  description = FALSE, ...)

Arguments

x
An object of class summary.simsum.

sstat
Summary statistics to include; can be a scalar value or a vector. Possible choices are:

all, all the summary statistics are returned This is the default option.

nsim, the number of replications with non-missing point estimates and standard error.

thetamean, average point estimate.

thetamedian, median point estimate.

se2mean, average standard error.

se2median, median standard error.

bias, bias in point estimate.

empse, empirical standard error.

mse, mean squared error.

relprec, percentage gain in precision relative to the reference method.

modelse, model-based standard error.

relerror, relative percentage error in standard error.

cover, coverage of a nominal level% confidence interval.

bccover, bias corrected coverage of a nominal level% confidence interval.

power, power of a (1 - level)% level test.#'

description
Append a column to the returned data.frame with a description of each summary statistic? Defaults to FALSE.

...
Ignored.

Value

A data.frame containing summary statistics from a simulation study.

See also

Examples

data(MIsim)
x <- simsum(
  data = MIsim, estvarname = "b", true = 0.5, se = "se",
  methodvar = "method", mcse = TRUE
)

#> `ref` was not specified, CC set as the reference

xs <- summary(x)
get_data(xs)

#>           stat           est         mcse  method        lower       upper
#> 1         nsim  1.000000e+03           NA      CC           NA          NA
#> 2    thetamean  5.167662e-01           NA      CC           NA          NA
#> 3  thetamedian  5.069935e-01           NA      CC           NA          NA
#> 4      se2mean  2.163731e-02           NA      CC           NA          NA
#> 5    se2median  2.114245e-02           NA      CC           NA          NA
#> 6         bias  1.676616e-02 0.0047786757      CC  0.007400129 0.026132193
#> 7        empse  1.511150e-01 0.0033807248      CC  0.144488895 0.157741093
#> 8          mse  2.309401e-02 0.0011338389      CC  0.020871727 0.025316293
#> 9      relprec  1.000000e+00 0.0000000000      CC  1.000000000 1.000000000
#> 10     modelse  1.470963e-01 0.0005274099      CC  0.146062561 0.148129970
#> 11    relerror -2.659384e+00 2.2049438356      CC -6.980994702 1.662226309
#> 12       cover  9.430000e-01 0.0073315073      CC  0.928630510 0.957369490
#> 13     bccover  9.400000e-01 0.0075099933      CC  0.925280684 0.954719316
#> 14       power  9.460000e-01 0.0071473072      CC  0.931991535 0.960008465
#> 15        nsim  1.000000e+03           NA MI_LOGT           NA          NA
#> 16   thetamean  5.009231e-01           NA MI_LOGT           NA          NA
#> 17 thetamedian  4.969223e-01           NA MI_LOGT           NA          NA
#> 18     se2mean  1.820915e-02           NA MI_LOGT           NA          NA
#> 19   se2median  1.721574e-02           NA MI_LOGT           NA          NA
#> 20        bias  9.230987e-04 0.0041744101 MI_LOGT -0.007258595 0.009104792
#> 21       empse  1.320064e-01 0.0029532306 MI_LOGT  0.126218211 0.137794663
#> 22         mse  1.740913e-02 0.0008812805 MI_LOGT  0.015681848 0.019136404
#> 23     relprec  1.310463e+00 0.0393747264 MI_LOGT  1.233290364 1.387636456
#> 24     modelse  1.349413e-01 0.0006046041 MI_LOGT  0.133756280 0.136126285
#> 25    relerror  2.223259e+00 2.3317772667 MI_LOGT -2.346940122 6.793458804
#> 26       cover  9.490000e-01 0.0069569390 MI_LOGT  0.935364650 0.962635350
#> 27     bccover  9.490000e-01 0.0069569390 MI_LOGT  0.935364650 0.962635350
#> 28       power  9.690000e-01 0.0054807846 MI_LOGT  0.958257860 0.979742140
#> 29        nsim  1.000000e+03           NA    MI_T           NA          NA
#> 30   thetamean  4.988092e-01           NA    MI_T           NA          NA
#> 31 thetamedian  4.939111e-01           NA    MI_T           NA          NA
#> 32     se2mean  1.791169e-02           NA    MI_T           NA          NA
#> 33   se2median  1.693191e-02           NA    MI_T           NA          NA
#> 34        bias -1.190835e-03 0.0042509767    MI_T -0.009522596 0.007140926
#> 35       empse  1.344277e-01 0.0030073985    MI_T  0.128533294 0.140322080
#> 36         mse  1.805415e-02 0.0009112249    MI_T  0.016268182 0.019840118
#> 37     relprec  1.263682e+00 0.0384237911    MI_T  1.188372366 1.338990859
#> 38     modelse  1.338346e-01 0.0005856362    MI_T  0.132686735 0.134982387
#> 39    relerror -4.412233e-01 2.2689747610    MI_T -4.888332074 4.005885552
#> 40       cover  9.430000e-01 0.0073315073    MI_T  0.928630510 0.957369490
#> 41     bccover  9.430000e-01 0.0073315073    MI_T  0.928630510 0.957369490
#> 42       power  9.630000e-01 0.0059691708    MI_T  0.951300640 0.974699360

# Exporting only bias and coverage:
get_data(xs, sstat = c("bias", "cover"))

#>     stat           est        mcse  method        lower       upper
#> 6   bias  0.0167661608 0.004778676      CC  0.007400129 0.026132193
#> 12 cover  0.9430000000 0.007331507      CC  0.928630510 0.957369490
#> 20  bias  0.0009230987 0.004174410 MI_LOGT -0.007258595 0.009104792
#> 26 cover  0.9490000000 0.006956939 MI_LOGT  0.935364650 0.962635350
#> 34  bias -0.0011908351 0.004250977    MI_T -0.009522596 0.007140926
#> 40 cover  0.9430000000 0.007331507    MI_T  0.928630510 0.957369490

# Append a column with a description of each summary statistic exported:
get_data(xs, sstat = c("bias", "cover"), description = TRUE)

#>    stat           est        mcse  method        lower       upper
#> 1  bias  0.0167661608 0.004778676      CC  0.007400129 0.026132193
#> 2  bias  0.0009230987 0.004174410 MI_LOGT -0.007258595 0.009104792
#> 3  bias -0.0011908351 0.004250977    MI_T -0.009522596 0.007140926
#> 4 cover  0.9430000000 0.007331507      CC  0.928630510 0.957369490
#> 5 cover  0.9490000000 0.006956939 MI_LOGT  0.935364650 0.962635350
#> 6 cover  0.9430000000 0.007331507    MI_T  0.928630510 0.957369490
#>                  description
#> 1     Bias in point estimate
#> 2     Bias in point estimate
#> 3     Bias in point estimate
#> 4 Coverage of nominal 95% CI
#> 5 Coverage of nominal 95% CI
#> 6 Coverage of nominal 95% CI