rsimsum: Analysis of Simulation Studies Including Monte Carlo Error

Summarise results from simulation studies and compute Monte Carlo standard errors of commonly used summary statistics. This package is modelled on the 'simsum' user-written command in 'Stata' (White I.R., 2010 https://www.stata-journal.com/article.html?article=st0200), further extending it with additional performance measures and functionality.

See also

Author

Maintainer: Alessandro Gasparini alessandro@ellessenne.xyz (ORCID)

Authors:

Ian R. White