Example of a simulation study on missing data

A dataset from a simulation study comparing different ways to handle missing covariates when fitting a Cox model (White and Royston, 2009). One thousand datasets were simulated, each containing normally distributed covariates \(x\) and \(z\) and time-to-event outcome. Both covariates has 20\

MIsim

Format

A data frame with 3,000 rows and 4 variables:

dataset Simulated dataset number.

method Method used (CC, MI_LOGT or MI_T).

b Point estimate.

se Standard error of the point estimate.

References

White, I.R., and P. Royston. 2009. Imputing missing covariate values for the Cox model. Statistics in Medicine 28(15):1982-1998 doi: 10.1002/sim.3618

Little, R.J.A., and D.B. Rubin. 2002. Statistical analysis with missing data. 2nd ed. Hoboken, NJ: Wiley doi: 10.1002/9781119013563

Examples

data("MIsim", package = "rsimsum")