print.summary.multisimsum

Print method for summary.multisimsum objects

# S3 method for summary.multisimsum
print(x, digits = 4, sstat = "all", ...)

Arguments

x
An object of class summary.multisimsum.

digits
Number of significant digits used for printing. Defaults to 4.

sstat
Summary statistics to print; can be a scalar value or a vector (for printing multiple summary statistics at once). Possible choices are:

all, all the summary statistics are printed. This is the default option.

nsim, the number of replications with non-missing point estimates and standard error.

thetamean, average point estimate.

thetamedian, median point estimate.

se2mean, average standard error.

se2median, median standard error.

bias, bias in point estimate.

empse, empirical standard error.

mse, mean squared error.

relprec, percentage gain in precision relative to the reference method.

modelse, model-based standard error.

relerror, relative percentage error in standard error.

cover, coverage of a nominal level% confidence interval.

bccover, bias corrected coverage of a nominal level% confidence interval.

power, power of a (1 - level)% level test.

...
Ignored.

Note

If sstat is a vector that contains all, all summary statistics are printed by default.

Examples

data(frailty)
ms <- multisimsum(
  data = frailty, par = "par", true = c(
    trt = -0.50,
    fv = 0.75
  ), estvarname = "b", se = "se", methodvar = "model",
  by = "fv_dist"
)

#> `ref` was not specified, Cox, Gamma set as the reference

sms <- summary(ms)
sms

#> 
#> Call:
#> 	multisimsum(data = frailty, par = "par", true = c(trt = -0.5, 
#>     fv = 0.75), estvarname = "b", se = "se", methodvar = "model", 
#>     by = "fv_dist")
#> 
#> Estimands variable: par 
#> 	Unique estimands: fv, trt 
#> 	True values: fv = 0.75, trt = -0.5 
#> 
#> Method variable: model 
#> 	Unique methods: Cox, Gamma, Cox, Log-Normal, RP(P), Gamma, RP(P), Log-Normal 
#> 	Reference method: Cox, Gamma 
#> 
#> By factors: fv_dist 
#> 
#>  -------------------------------------------------------------------------------- 
#> 
#> Summary statistics for estimand 'fv':
#> 
#> 	Method = Cox, Gamma, fv_dist = Gamma 
#>                                                    Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 976.0000     NA         NA
#>                             Average point estimate   0.7376     NA         NA
#>                              Median point estimate   0.7271     NA         NA
#>                             Average standard error   0.0203     NA         NA
#>                              Median standard error   0.0193     NA         NA
#>                             Bias in point estimate  -0.0124 0.0045    -0.0213
#>                           Empirical standard error   0.1421 0.0032     0.1358
#>                                 Mean squared error   0.0203 0.0010     0.0184
#>  % gain in precision relative to method Cox, Gamma   1.0000 0.0000     1.0000
#>                         Model-based standard error   0.1426 0.0008     0.1410
#>                 Relative % error in standard error   0.3574 2.3458    -4.2402
#>                         Coverage of nominal 95% CI   0.9201 0.0087     0.9031
#>          Bias corrected coverage of nominal 95% CI   0.9334 0.0080     0.9178
#>                             Power of 5% level test   1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>      -0.0035
#>       0.1484
#>       0.0222
#>       1.0000
#>       0.1442
#>       4.9550
#>       0.9371
#>       0.9490
#>       1.0000
#> 
#> 	Method = Cox, Log-Normal, fv_dist = Gamma 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate    0.9799     NA         NA
#>                              Median point estimate    0.9566     NA         NA
#>                             Average standard error    0.0600     NA         NA
#>                              Median standard error    0.0483     NA         NA
#>                             Bias in point estimate    0.2299 0.0076     0.2150
#>                           Empirical standard error    0.2406 0.0054     0.2301
#>                                 Mean squared error    0.1107 0.0055     0.1000
#>  % gain in precision relative to method Cox, Gamma    0.3487 0.0061     0.3367
#>                         Model-based standard error    0.2449 0.0027     0.2396
#>                 Relative % error in standard error    1.7896 2.5444    -3.1973
#>                         Coverage of nominal 95% CI    0.9220 0.0085     0.9054
#>          Bias corrected coverage of nominal 95% CI    0.8980 0.0096     0.8792
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.2448
#>       0.2512
#>       0.1214
#>       0.3608
#>       0.2503
#>       6.7765
#>       0.9386
#>       0.9168
#>       1.0000
#> 
#> 	Method = RP(P), Gamma, fv_dist = Gamma 
#>                                                    Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 971.0000     NA         NA
#>                             Average point estimate   0.7321     NA         NA
#>                              Median point estimate   0.7225     NA         NA
#>                             Average standard error   0.0202     NA         NA
#>                              Median standard error   0.0191     NA         NA
#>                             Bias in point estimate  -0.0179 0.0044    -0.0266
#>                           Empirical standard error   0.1387 0.0031     0.1325
#>                                 Mean squared error   0.0195 0.0009     0.0178
#>  % gain in precision relative to method Cox, Gamma   1.0500 0.0005     1.0491
#>                         Model-based standard error   0.1420 0.0008     0.1404
#>                 Relative % error in standard error   2.3922 2.3945    -2.3009
#>                         Coverage of nominal 95% CI   0.9300 0.0082     0.9139
#>          Bias corrected coverage of nominal 95% CI   0.9434 0.0074     0.9288
#>                             Power of 5% level test   1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>      -0.0092
#>       0.1448
#>       0.0213
#>       1.0510
#>       0.1435
#>       7.0852
#>       0.9460
#>       0.9579
#>       1.0000
#> 
#> 	Method = RP(P), Log-Normal, fv_dist = Gamma 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate    0.9847     NA         NA
#>                              Median point estimate    0.9597     NA         NA
#>                             Average standard error    0.0498     NA         NA
#>                              Median standard error    0.0442     NA         NA
#>                             Bias in point estimate    0.2347 0.0077     0.2196
#>                           Empirical standard error    0.2438 0.0055     0.2331
#>                                 Mean squared error    0.1145 0.0057     0.1033
#>  % gain in precision relative to method Cox, Gamma    0.3397 0.0059     0.3281
#>                         Model-based standard error    0.2232 0.0019     0.2195
#>                 Relative % error in standard error   -8.4531 2.1885   -12.7425
#>                         Coverage of nominal 95% CI    0.9030 0.0094     0.8847
#>          Bias corrected coverage of nominal 95% CI    0.8930 0.0098     0.8738
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.2499
#>       0.2545
#>       0.1256
#>       0.3512
#>       0.2269
#>      -4.1636
#>       0.9213
#>       0.9122
#>       1.0000
#> 
#> 	Method = Cox, Gamma, fv_dist = Log-Normal 
#>                                                    Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 957.0000     NA         NA
#>                             Average point estimate   0.6436     NA         NA
#>                              Median point estimate   0.6365     NA         NA
#>                             Average standard error   0.0156     NA         NA
#>                              Median standard error   0.0149     NA         NA
#>                             Bias in point estimate  -0.1064 0.0043    -0.1148
#>                           Empirical standard error   0.1320 0.0030     0.1261
#>                                 Mean squared error   0.0287 0.0010     0.0267
#>  % gain in precision relative to method Cox, Gamma   1.0000 0.0000     1.0000
#>                         Model-based standard error   0.1249 0.0008     0.1233
#>                 Relative % error in standard error  -5.3912 2.2447    -9.7907
#>                         Coverage of nominal 95% CI   0.7503 0.0140     0.7228
#>          Bias corrected coverage of nominal 95% CI   0.9164 0.0089     0.8989
#>                             Power of 5% level test   1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>      -0.0980
#>       0.1379
#>       0.0308
#>       1.0000
#>       0.1264
#>      -0.9918
#>       0.7777
#>       0.9339
#>       1.0000
#> 
#> 	Method = Cox, Log-Normal, fv_dist = Log-Normal 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate    0.7325     NA         NA
#>                              Median point estimate    0.7182     NA         NA
#>                             Average standard error    0.0230     NA         NA
#>                              Median standard error    0.0206     NA         NA
#>                             Bias in point estimate   -0.0175 0.0049    -0.0271
#>                           Empirical standard error    0.1554 0.0035     0.1485
#>                                 Mean squared error    0.0244 0.0011     0.0222
#>  % gain in precision relative to method Cox, Gamma    0.7217 0.0150     0.6922
#>                         Model-based standard error    0.1517 0.0013     0.1493
#>                 Relative % error in standard error   -2.3382 2.3296    -6.9040
#>                         Coverage of nominal 95% CI    0.9020 0.0094     0.8836
#>          Bias corrected coverage of nominal 95% CI    0.9130 0.0089     0.8955
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>      -0.0079
#>       0.1622
#>       0.0266
#>       0.7512
#>       0.1542
#>       2.2276
#>       0.9204
#>       0.9305
#>       1.0000
#> 
#> 	Method = RP(P), Gamma, fv_dist = Log-Normal 
#>                                                    Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 997.0000     NA         NA
#>                             Average point estimate   0.6434     NA         NA
#>                              Median point estimate   0.6324     NA         NA
#>                             Average standard error   0.0158     NA         NA
#>                              Median standard error   0.0149     NA         NA
#>                             Bias in point estimate  -0.1066 0.0041    -0.1147
#>                           Empirical standard error   0.1307 0.0029     0.1249
#>                                 Mean squared error   0.0284 0.0010     0.0264
#>  % gain in precision relative to method Cox, Gamma   1.0205 0.0005     1.0196
#>                         Model-based standard error   0.1258 0.0007     0.1244
#>                 Relative % error in standard error  -3.7112 2.2295    -8.0809
#>                         Coverage of nominal 95% CI   0.7683 0.0134     0.7421
#>          Bias corrected coverage of nominal 95% CI   0.9308 0.0080     0.9150
#>                             Power of 5% level test   1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>      -0.0985
#>       0.1364
#>       0.0304
#>       1.0214
#>       0.1272
#>       0.6586
#>       0.7945
#>       0.9465
#>       1.0000
#> 
#> 	Method = RP(P), Log-Normal, fv_dist = Log-Normal 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate    0.7348     NA         NA
#>                              Median point estimate    0.7199     NA         NA
#>                             Average standard error    0.0254     NA         NA
#>                              Median standard error    0.0235     NA         NA
#>                             Bias in point estimate   -0.0152 0.0050    -0.0250
#>                           Empirical standard error    0.1570 0.0035     0.1501
#>                                 Mean squared error    0.0248 0.0012     0.0226
#>  % gain in precision relative to method Cox, Gamma    0.7069 0.0146     0.6783
#>                         Model-based standard error    0.1594 0.0011     0.1573
#>                 Relative % error in standard error    1.5422 2.3708    -3.1044
#>                         Coverage of nominal 95% CI    0.9280 0.0082     0.9120
#>          Bias corrected coverage of nominal 95% CI    0.9360 0.0077     0.9208
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>      -0.0055
#>       0.1639
#>       0.0271
#>       0.7355
#>       0.1615
#>       6.1888
#>       0.9440
#>       0.9512
#>       1.0000
#> 
#>  -------------------------------------------------------------------------------- 
#> 
#> Summary statistics for estimand 'trt':
#> 
#> 	Method = Cox, Gamma, fv_dist = Gamma 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate   -0.5006     NA         NA
#>                              Median point estimate   -0.5011     NA         NA
#>                             Average standard error    0.0026     NA         NA
#>                              Median standard error    0.0026     NA         NA
#>                             Bias in point estimate   -0.0006 0.0016    -0.0037
#>                           Empirical standard error    0.0508 0.0011     0.0485
#>                                 Mean squared error    0.0026 0.0001     0.0023
#>  % gain in precision relative to method Cox, Gamma    1.0000 0.0000     1.0000
#>                         Model-based standard error    0.0506 0.0000     0.0506
#>                 Relative % error in standard error   -0.2017 2.2340    -4.5803
#>                         Coverage of nominal 95% CI    0.9500 0.0069     0.9365
#>          Bias corrected coverage of nominal 95% CI    0.9500 0.0069     0.9365
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0026
#>       0.0530
#>       0.0028
#>       1.0000
#>       0.0507
#>       4.1769
#>       0.9635
#>       0.9635
#>       1.0000
#> 
#> 	Method = Cox, Log-Normal, fv_dist = Gamma 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate   -0.5013     NA         NA
#>                              Median point estimate   -0.5021     NA         NA
#>                             Average standard error    0.0026     NA         NA
#>                              Median standard error    0.0026     NA         NA
#>                             Bias in point estimate   -0.0013 0.0016    -0.0045
#>                           Empirical standard error    0.0509 0.0011     0.0486
#>                                 Mean squared error    0.0026 0.0001     0.0023
#>  % gain in precision relative to method Cox, Gamma    0.9955 0.0011     0.9933
#>                         Model-based standard error    0.0507 0.0000     0.0506
#>                 Relative % error in standard error   -0.3890 2.2298    -4.7594
#>                         Coverage of nominal 95% CI    0.9490 0.0070     0.9354
#>          Bias corrected coverage of nominal 95% CI    0.9500 0.0069     0.9365
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0018
#>       0.0531
#>       0.0028
#>       0.9978
#>       0.0508
#>       3.9813
#>       0.9626
#>       0.9635
#>       1.0000
#> 
#> 	Method = RP(P), Gamma, fv_dist = Gamma 
#>                                                    Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 971.0000     NA         NA
#>                             Average point estimate  -0.5003     NA         NA
#>                              Median point estimate  -0.5010     NA         NA
#>                             Average standard error   0.0026     NA         NA
#>                              Median standard error   0.0026     NA         NA
#>                             Bias in point estimate  -0.0003 0.0016    -0.0035
#>                           Empirical standard error   0.0506 0.0011     0.0484
#>                                 Mean squared error   0.0026 0.0001     0.0023
#>  % gain in precision relative to method Cox, Gamma   1.0044 0.0008     1.0027
#>                         Model-based standard error   0.0506 0.0000     0.0505
#>                 Relative % error in standard error  -0.0544 2.2704    -4.5043
#>                         Coverage of nominal 95% CI   0.9506 0.0070     0.9369
#>          Bias corrected coverage of nominal 95% CI   0.9506 0.0070     0.9369
#>                             Power of 5% level test   1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0029
#>       0.0529
#>       0.0028
#>       1.0061
#>       0.0507
#>       4.3955
#>       0.9642
#>       0.9642
#>       1.0000
#> 
#> 	Method = RP(P), Log-Normal, fv_dist = Gamma 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate   -0.5015     NA         NA
#>                              Median point estimate   -0.5025     NA         NA
#>                             Average standard error    0.0026     NA         NA
#>                              Median standard error    0.0026     NA         NA
#>                             Bias in point estimate   -0.0015 0.0016    -0.0047
#>                           Empirical standard error    0.0509 0.0011     0.0487
#>                                 Mean squared error    0.0026 0.0001     0.0023
#>  % gain in precision relative to method Cox, Gamma    0.9942 0.0016     0.9910
#>                         Model-based standard error    0.0507 0.0000     0.0506
#>                 Relative % error in standard error   -0.4330 2.2288    -4.8014
#>                         Coverage of nominal 95% CI    0.9500 0.0069     0.9365
#>          Bias corrected coverage of nominal 95% CI    0.9490 0.0070     0.9354
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0016
#>       0.0531
#>       0.0028
#>       0.9974
#>       0.0508
#>       3.9355
#>       0.9635
#>       0.9626
#>       1.0000
#> 
#> 	Method = Cox, Gamma, fv_dist = Log-Normal 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate   -0.5006     NA         NA
#>                              Median point estimate   -0.5014     NA         NA
#>                             Average standard error    0.0023     NA         NA
#>                              Median standard error    0.0022     NA         NA
#>                             Bias in point estimate   -0.0006 0.0015    -0.0035
#>                           Empirical standard error    0.0474 0.0011     0.0453
#>                                 Mean squared error    0.0022 0.0001     0.0020
#>  % gain in precision relative to method Cox, Gamma    1.0000 0.0000     1.0000
#>                         Model-based standard error    0.0475 0.0000     0.0474
#>                 Relative % error in standard error    0.2507 2.2433    -4.1461
#>                         Coverage of nominal 95% CI    0.9410 0.0075     0.9264
#>          Bias corrected coverage of nominal 95% CI    0.9420 0.0074     0.9275
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0024
#>       0.0494
#>       0.0024
#>       1.0000
#>       0.0475
#>       4.6474
#>       0.9556
#>       0.9565
#>       1.0000
#> 
#> 	Method = Cox, Log-Normal, fv_dist = Log-Normal 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate   -0.5014     NA         NA
#>                              Median point estimate   -0.5021     NA         NA
#>                             Average standard error    0.0023     NA         NA
#>                              Median standard error    0.0022     NA         NA
#>                             Bias in point estimate   -0.0014 0.0015    -0.0044
#>                           Empirical standard error    0.0474 0.0011     0.0453
#>                                 Mean squared error    0.0022 0.0001     0.0020
#>  % gain in precision relative to method Cox, Gamma    0.9986 0.0014     0.9959
#>                         Model-based standard error    0.0475 0.0000     0.0474
#>                 Relative % error in standard error    0.1815 2.2417    -4.2123
#>                         Coverage of nominal 95% CI    0.9420 0.0074     0.9275
#>          Bias corrected coverage of nominal 95% CI    0.9400 0.0075     0.9253
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0015
#>       0.0495
#>       0.0024
#>       1.0013
#>       0.0475
#>       4.5752
#>       0.9565
#>       0.9547
#>       1.0000
#> 
#> 	Method = RP(P), Gamma, fv_dist = Log-Normal 
#>                                                    Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 997.0000     NA         NA
#>                             Average point estimate  -0.5006     NA         NA
#>                              Median point estimate  -0.5014     NA         NA
#>                             Average standard error   0.0023     NA         NA
#>                              Median standard error   0.0022     NA         NA
#>                             Bias in point estimate  -0.0006 0.0015    -0.0036
#>                           Empirical standard error   0.0473 0.0011     0.0453
#>                                 Mean squared error   0.0022 0.0001     0.0020
#>  % gain in precision relative to method Cox, Gamma   1.0009 0.0009     0.9992
#>                         Model-based standard error   0.0475 0.0000     0.0474
#>                 Relative % error in standard error   0.3319 2.2485    -4.0750
#>                         Coverage of nominal 95% CI   0.9428 0.0074     0.9284
#>          Bias corrected coverage of nominal 95% CI   0.9428 0.0074     0.9284
#>                             Power of 5% level test   1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0023
#>       0.0494
#>       0.0024
#>       1.0026
#>       0.0476
#>       4.7389
#>       0.9572
#>       0.9572
#>       1.0000
#> 
#> 	Method = RP(P), Log-Normal, fv_dist = Log-Normal 
#>                                                     Estimate   MCSE Lower 2.5%
#>         Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                             Average point estimate   -0.5016     NA         NA
#>                              Median point estimate   -0.5022     NA         NA
#>                             Average standard error    0.0023     NA         NA
#>                              Median standard error    0.0022     NA         NA
#>                             Bias in point estimate   -0.0016 0.0015    -0.0045
#>                           Empirical standard error    0.0474 0.0011     0.0453
#>                                 Mean squared error    0.0022 0.0001     0.0020
#>  % gain in precision relative to method Cox, Gamma    0.9979 0.0016     0.9948
#>                         Model-based standard error    0.0475 0.0000     0.0474
#>                 Relative % error in standard error    0.2101 2.2424    -4.1849
#>                         Coverage of nominal 95% CI    0.9430 0.0073     0.9286
#>          Bias corrected coverage of nominal 95% CI    0.9410 0.0075     0.9264
#>                             Power of 5% level test    1.0000 0.0000     1.0000
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0013
#>       0.0495
#>       0.0024
#>       1.0010
#>       0.0476
#>       4.6051
#>       0.9574
#>       0.9556
#>       1.0000