Summarising simsum objects

The summary() method for objects of class simsum returns confidence intervals for performance measures based on Monte Carlo standard errors.

# S3 method for simsum
summary(object, ci_level = 0.95, ...)

Arguments

object
An object of class simsum.

ci_level
Significance level for confidence intervals based on Monte Carlo standard errors. Ignored if a simsum object with mcse = FALSE is passed.

...
Ignored.

Value

An object of class summary.simsum.

Note

Confidence intervals are constructed using quantiles from a normal distribution with the specified ci_level confidence level, and are therefore symmetric by definition.

See also

Examples

data("MIsim")
object <- simsum(
  data = MIsim, estvarname = "b", true = 0.5, se = "se",
  methodvar = "method", mcse = TRUE
)

#> `ref` was not specified, CC set as the reference

xs <- summary(object)
xs

#> 
#> Call:
#> 	simsum(data = MIsim, estvarname = "b", true = 0.5, se = "se", 
#>     methodvar = "method", mcse = TRUE)
#> 
#> Method variable: method 
#> 	Unique methods: CC, MI_LOGT, MI_T 
#> 	Reference method: CC 
#> By factors: none
#> 
#> Summary statistics:
#> 
#> 	Method = CC 
#>                                              Estimate   MCSE Lower 2.5%
#>  Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                      Average point estimate    0.5168     NA         NA
#>                       Median point estimate    0.5070     NA         NA
#>                      Average standard error    0.0216     NA         NA
#>                       Median standard error    0.0211     NA         NA
#>                      Bias in point estimate    0.0168 0.0048     0.0074
#>                    Empirical standard error    0.1511 0.0034     0.1445
#>                          Mean squared error    0.0231 0.0011     0.0209
#>   % gain in precision relative to method CC    1.0000 0.0000     1.0000
#>                  Model-based standard error    0.1471 0.0005     0.1461
#>          Relative % error in standard error   -2.6594 2.2049    -6.9810
#>                  Coverage of nominal 95% CI    0.9430 0.0073     0.9286
#>   Bias corrected coverage of nominal 95% CI    0.9400 0.0075     0.9253
#>                      Power of 5% level test    0.9460 0.0071     0.9320
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0261
#>       0.1577
#>       0.0253
#>       1.0000
#>       0.1481
#>       1.6622
#>       0.9574
#>       0.9547
#>       0.9600
#> 
#> 	Method = MI_LOGT 
#>                                              Estimate   MCSE Lower 2.5%
#>  Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                      Average point estimate    0.5009     NA         NA
#>                       Median point estimate    0.4969     NA         NA
#>                      Average standard error    0.0182     NA         NA
#>                       Median standard error    0.0172     NA         NA
#>                      Bias in point estimate    0.0009 0.0042    -0.0073
#>                    Empirical standard error    0.1320 0.0030     0.1262
#>                          Mean squared error    0.0174 0.0009     0.0157
#>   % gain in precision relative to method CC    1.3105 0.0394     1.2333
#>                  Model-based standard error    0.1349 0.0006     0.1338
#>          Relative % error in standard error    2.2233 2.3318    -2.3469
#>                  Coverage of nominal 95% CI    0.9490 0.0070     0.9354
#>   Bias corrected coverage of nominal 95% CI    0.9490 0.0070     0.9354
#>                      Power of 5% level test    0.9690 0.0055     0.9583
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0091
#>       0.1378
#>       0.0191
#>       1.3876
#>       0.1361
#>       6.7935
#>       0.9626
#>       0.9626
#>       0.9797
#> 
#> 	Method = MI_T 
#>                                              Estimate   MCSE Lower 2.5%
#>  Simulations with non-missing estimates/SEs 1000.0000     NA         NA
#>                      Average point estimate    0.4988     NA         NA
#>                       Median point estimate    0.4939     NA         NA
#>                      Average standard error    0.0179     NA         NA
#>                       Median standard error    0.0169     NA         NA
#>                      Bias in point estimate   -0.0012 0.0043    -0.0095
#>                    Empirical standard error    0.1344 0.0030     0.1285
#>                          Mean squared error    0.0181 0.0009     0.0163
#>   % gain in precision relative to method CC    1.2637 0.0384     1.1884
#>                  Model-based standard error    0.1338 0.0006     0.1327
#>          Relative % error in standard error   -0.4412 2.2690    -4.8883
#>                  Coverage of nominal 95% CI    0.9430 0.0073     0.9286
#>   Bias corrected coverage of nominal 95% CI    0.9430 0.0073     0.9286
#>                      Power of 5% level test    0.9630 0.0060     0.9513
#>  Upper 97.5%
#>           NA
#>           NA
#>           NA
#>           NA
#>           NA
#>       0.0071
#>       0.1403
#>       0.0198
#>       1.3390
#>       0.1350
#>       4.0059
#>       0.9574
#>       0.9574
#>       0.9747